+++ app/src/main/java/com/apptive/braini/presentation/viewmodel/LoginViewModel.kt
@@ -11,6 +11,7 @@ import com.apptive.braini.domain.log
 import com.apptive.braini.domain.model.Account
 import com.apptive.braini.domain.utils.AccountManager
 import com.apptive.braini.domain.utils.GoogleOneTap
+import com.apptive.braini.domain.utils.KakaoLogin
 import com.apptive.braini.presentation.viewmodel.interfaces.ILoginViewModel
 import com.kakao.sdk.auth.AuthApiClient
 import com.kakao.sdk.common.model.ClientError
@@ -32,51 +33,7 @@ class LoginViewModel @Inject constructor(
     override fun isLoggedIn(): Boolean = accountManager.isLoggedIn()
 
     override fun loginWithKakao(context: Context, navigate: () -> Unit) {
-        // 토큰이 유효한지 체크한다.
-        tryLoginWithKakaoToken(
-            callLoginApi = {
-                callKakaoLoginApi(context, navigate)
-            },
-            onSuccess = {
-                loginWithCachedKakaoAccount(navigate)
-            },
-        )
-    }
-
-    private fun tryLoginWithKakaoToken(
-        onSuccess: () -> Unit,
-        callLoginApi: () -> Unit,
-    ) {
-        if (!AuthApiClient.instance.hasToken()) return callLoginApi()
-
-        UserApiClient.instance.accessTokenInfo { tokenInfo, error ->
-            if (error == null) return@accessTokenInfo onSuccess()
-            if (error is KakaoSdkError && error.isInvalidTokenError()) callLoginApi()
-
-            // 기타 에러
-            "카카오 로그인 에러! $error".log()
-        }
-    }
-
-    private fun callKakaoLoginApi(context: Context, navigate: () -> Unit) {
-        // 이곳에서 사용하는 accessToken은 REST API에 이용된다.
-        UserApiClient.instance.loginWithKakaoTalk(context) { token, error ->
-            if (error != null) handleErrorWithKakao(context, error)
-            else if (token != null) loginWithCachedKakaoAccount(navigate)
-        }
-    }
-
-    private fun loginWithCachedKakaoAccount(navigate: () -> Unit) {
-        UserApiClient.instance.me { user, error ->
-            val id = user?.id ?: 0L
-            val email = user?.kakaoAccount?.email ?: ""
-            val name = user?.kakaoAccount?.profile?.nickname ?: ""
-
-            val account = Account(
-                id = id.toString(),
-                email = email,
-                name = name
-            )
+        KakaoLogin.login(context) { account ->
             accountManager.login(account)
             navigate()
         }
@@ -94,19 +51,5 @@ class LoginViewModel @Inject constructor(
 
     override fun googleSignIn(launcher: IntentSenderLauncher) {
         googleSignIn.beginSignIn(launcher = launcher)
-    }
-
-    private fun handleErrorWithKakao(context: Context, error: Throwable?) {
-        "로그인 실패 $error".log()
-        if (error is ClientError) handleClientErrorWithKakao(context, error)
-    }
-
-    private fun handleClientErrorWithKakao(context: Context, error: ClientError) {
-        when(error.reason) {
-            ClientErrorCause.NotSupported -> {
-                Toast.makeText(context, "카카오톡이 설치되지 않았습니다.", Toast.LENGTH_LONG).show()
-            }
-            else -> {}
-        }
     }
 }
